requirements.txt
flask
flask-wtf
transformers
sentence-transformers
pdfminer.six
scikit-learn
spacy
python-dotenv 
gunicorn 